fix: Remove baseUrl in favor of authorizeUrl and tokenUrl
I made a wrong assumption about baseUrl always being the same on all providers when sending a user
to the authorize url and querying the token endpoint. But that was wrong. Now the endpoints has tp
be defined explicitly.

BREAKING CHANGE: Remove props baseUrl, authorizeEndpoint and tokenEndpoint in favor of more explicit
authorizeUrl and tokenUrl

fix #4